% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dimpler.R
\name{dimpler}
\alias{dimpler}
\title{Creates a new Datatbl from a dataframe}
\usage{
dimpler(data, name, width = NULL, height = NULL)
}
\arguments{
\item{data}{is an object of class data.frame or Tblpkg}
}
\value{
dimple visualizations from rcdimple package
}
\description{
Graphs an input table a given dimple preset.
}
\examples{
\dontrun{

}
}

